Resolved issue in pypresence's utils.py preventing Discord status setup while logged in as root user.
Due to the setup.sh file creating services to execute htb-presence.py as sudo, an issue arose because the utils.py script in the pypresence library checked for Discord's Unix Socket in /run/user/{linux_user_guid} or XDG_RUNTIME_DIR to verify if Discord was running. When the library was executed as sudo, it searched in the sudo user's XDG_RUNTIME_DIR, where Discord wasn't running, causing the Rich Presence instance to fail.

utils.py was modified to ensure it still finds the Discord Unix Socket even when executed as sudo.